gemini-folder/pyproject.toml
[project]
name = "gemini-folder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=44.0.1",
    "google-genai>=1.2.0",
    "pypdf>=5.3.0",
    "tqdm>=4.67.1",
]
